Reorg and complete api surface of confirmSignup
