+++ src/Serde.bs.js
@@ -1,0 +1,99 @@
+// Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
+'use strict';
+
+var $$Array = require("bs-platform/lib/js/array.js");
+var Js_dict = require("bs-platform/lib/js/js_dict.js");
+var Js_json = require("bs-platform/lib/js/js_json.js");
+var Caml_option = require("bs-platform/lib/js/caml_option.js");
+
+function parseCognitoError(err) {
+  var match = Js_json.decodeObject(err);
+  if (match !== undefined) {
+    var envelope = Caml_option.valFromOption(match);
+    var match$1 = Js_dict.get(envelope, "__type");
+    var match$2 = Js_dict.get(envelope, "message");
+    if (match$1 !== undefined && match$2 !== undefined) {
+      var match$3 = Js_json.decodeString(Caml_option.valFromOption(match$1));
+      var match$4 = Js_json.decodeString(Caml_option.valFromOption(match$2));
+      if (match$3 !== undefined && match$4 !== undefined) {
+        return /* record */[
+                /* __type */match$3,
+                /* message */match$4
+              ];
+      } else {
+        return ;
+      }
+    } else {
+      return ;
+    }
+  }
+  
+}
+
+function makeSignupResponse(json) {
+  var match = Js_json.decodeObject(json);
+  if (match !== undefined) {
+    var env = Caml_option.valFromOption(match);
+    var match$1 = Js_dict.get(env, "UserConfirmed");
+    var match$2 = Js_dict.get(env, "UserSub");
+    var match$3 = Js_dict.get(env, "CodeDeliveryDetails");
+    if (match$1 !== undefined && match$2 !== undefined && match$3 !== undefined) {
+      var match$4 = Js_json.decodeBoolean(Caml_option.valFromOption(match$1));
+      var match$5 = Js_json.decodeString(Caml_option.valFromOption(match$2));
+      var match$6 = Js_json.decodeObject(Caml_option.valFromOption(match$3));
+      if (match$4 !== undefined && match$5 !== undefined && match$6 !== undefined) {
+        var delivery = Caml_option.valFromOption(match$6);
+        var match$7 = Js_dict.get(delivery, "AttributeName");
+        var match$8 = Js_dict.get(delivery, "DeliveryMedium");
+        var match$9 = Js_dict.get(delivery, "Destination");
+        if (match$7 !== undefined && match$8 !== undefined && match$9 !== undefined) {
+          var match$10 = Js_json.decodeString(Caml_option.valFromOption(match$7));
+          var match$11 = Js_json.decodeString(Caml_option.valFromOption(match$9));
+          var match$12 = Js_json.decodeString(Caml_option.valFromOption(match$8));
+          var match$13;
+          if (match$12 !== undefined) {
+            var medium = match$12;
+            match$13 = medium === "EMAIL" ? /* Email */0 : (
+                medium === "SMS" ? /* SMS */1 : /* UnknownDeliveryMedium */2
+              );
+          } else {
+            match$13 = /* UnknownDeliveryMedium */2;
+          }
+          if (match$10 !== undefined && match$11 !== undefined) {
+            return /* record */[
+                    /* codeDeliveryDetails : record */[
+                      /* attributeName */match$10,
+                      /* deliveryMedium */match$13,
+                      /* destination */match$11
+                    ],
+                    /* userConfirmed */match$4,
+                    /* userSub */match$5
+                  ];
+          } else {
+            return ;
+          }
+        } else {
+          return ;
+        }
+      } else {
+        return ;
+      }
+    } else {
+      return ;
+    }
+  }
+  
+}
+
+function jsonMapString(arr) {
+  return $$Array.map((function (item) {
+                return Js_dict.map((function (value) {
+                              return value;
+                            }), item);
+              }), arr);
+}
+
+exports.parseCognitoError = parseCognitoError;
+exports.makeSignupResponse = makeSignupResponse;
+exports.jsonMapString = jsonMapString;
+/* No side effect */
