+++ types.json
@@ -1,0 +1,20 @@
+{
+  "milkSchemaVersion": 2,
+  "version": 1,
+  "engines": {
+    "bs.json": {
+      "output": "./src/Serde.re"
+    }
+  },
+  "entries": [
+    {
+      "file": "./src/Types.re",
+      "type": "config"
+    },
+    {
+      "file": "./src/Node.re",
+      "type": "t",
+      "publicName": "node"
+    }
+  ]
+}
